Move input.input to abc.input
