+++ logprep/connector/confluent_kafka/input.py
@@ -16,7 +16,7 @@ from logprep.connector.confluent_kafka.common import (
     ConfluentKafkaFactory,
     UnknownOptionError,
 )
-from logprep.input.input import Input, CriticalInputError
+from logprep.abc.input import Input, CriticalInputError
 from logprep.output.output import Output
 from logprep.util.helper import add_field_to, get_dotted_field_value
