+++ logprep/connector/connector_factory.py
@@ -6,7 +6,7 @@ from logprep.connector.connector_factory_error import (
     UnknownConnectorTypeError,
     InvalidConfigurationError,
 )
-from logprep.input.input import Input
+from logprep.abc.input import Input
 from logprep.output.output import Output
 from logprep.connector.dummy.input import DummyInput
 from logprep.connector.jsonl.input import JsonlInput
